requirements.txt
# Core dependencies
streamlit>=1.25.0
langchain>=0.0.267
langchain-core>=0.1.4
langchain-community>=0.0.12
langgraph>=0.0.21

# Vector database
qdrant-client>=1.6.3

# Embedding models
sentence-transformers>=2.2.2
torch>=2.0.0

# Document processing
PyPDF2>=3.0.0
pdf2image>=1.16.3
docx2txt>=0.8
python-docx>=0.8.11

# LLM inference
transformers>=4.36.0
huggingface-hub>=0.17.3

# Utilities
python-dotenv>=1.0.0
uuid>=1.30
logging>=0.5.1

# Optional: for CUDA support
# nvidia-cuda-runtime-cu12>=12.0
# nvidia-cudnn-cu12>=8.9.0